bin_stream: clean code and update API
The __len__ cannot be used any more in bin_stream: Python returns an int object,
which will cap values to 0x7FFFFFFF on 32 bit systems. A binary can have a base
address higher than this, making it impossible to handle such programs.

The __len__ is replaced by a call to getlen.

Elfesteem needs an update to support 'max_addr' API in order to retrieve the
elf/pe max virtual address.