Support new prototype for tcc_add_file()
Since April, tcc_add_file() has taken a third parameter which forces
the type of input file.
As no versioning of the API is provided, we check for the presence of
the preprocessor macro we have to use with the new version of the
prototype so that we keep compile-time backward compatibility.
Note that updating libtcc.so without updating miasm will probably lead
to the jitter not working, but that's on libtcc's side: non-compatible
API change without major revision bump.