bugfix: don't use 128-bit masking constants (unhandled by gcc)
When translating ExprOp to C language, integer constants masks can be
huge, especially with multiplication and division operations. They
can even be wider than 64 bits, the maximum size gcc can handle. Thus
when an attempt to write such a mask is done, we simply do not write
it in the C code.