x86 semantics: better implementation of the cond-move template
As often, an unmet condition does not mean the destination operand
is not modified. For instance, with 32-bits operands in 64-bits
addressing mode, the cmov instruction modifies the destination
operand, even if the condition is not satisfied (and in particular
if the value hold by the register does not fit in a 32-bits value
before the instruction).

This modification solves the following failing test (amongst others):

[0000003] cmovneq    R=1234fdb512345678 (pc=406059)
Obtained: cmovnel    R=1234fdb512345678
Expected: cmovnel    R=0000000012345678