Renaming i386 qemu tests and preparing 64-bits ones
In test_all.py, the QEMU_TESTS list has been reverted back to a
dictionary to permit jitter blacklisting when x86_64 will be fully
available to testing. Indeed, some operations are not compatible with
certain types of jitters (for instance test_mul cannot be performed
by tcc, because it does not yet handle 128-bits integers properly).