Do not use exit(EXIT_FAILURE) for memory access errors
The use of exit(EXIT_FAILURE) was not soft enough. Memory errors are already
handled by other parts of Miasm, we do not want the taint analysis engine to
stop Miasm.
In some cases this was a real issue. For example, when working with SEH we
want to continue the analysis even if a memory access error occurs.