Fix the 011111 group of store instructions wrt r0
The 011111 group of store instructions is split in two sets: one in
which r0 has value 0, and another where r0 mustn't be used.
A "typographical inversion" interverted the two sets, leading to
a semantics error when r0 is used with the subset which expects a
hardcoded 0 value.

Found by @commial, thanks!