Include missing functions from llvm's compiler-rt into Jitllvm
Under Windows, we also need to compile these with clang (because MSVCRT doesn't seem to
have proper uint128_t support), so automatically check if clang is
installed (through the registry) and force distutils to use it (with a
not so nice hack).

Last but not least, this only works for 64 bit systems, so disable the mn_div
test if this is not the case.