Fixed no-op in example/expression/simplification_add
Previously, the test would not do anything, as the simplification added in the test is already a part of the default enabled simplifications:

```
Without adding the simplification:
	a + a + a = a * 0x3
After adding the simplification:
	a + a + a = a * 0x3
```

This also meant that editing the added simplification would have no effect (since the expression was already modified by the default simplifier, it would never match the custom one).

This commit replaces `expr_simp` by a newly-created `simp` that doesn't have any simplifications enabled, so the one added by the test has an impact.